HBX-1685: Move class 'org.hibernate.tool.hbm2x.visitor.DefaultValueVisitor' to 'org.hibernate.tool.internal.export.common.DefaultValueVisitor'
