+++ orm/src/main/java/org/hibernate/tool/internal/export/pojo/BasicPOJOClass.java
@@ -21,7 +21,7 @@ import org.hibernate.mapping.Selectable;
 import org.hibernate.mapping.Set;
 import org.hibernate.mapping.SimpleValue;
 import org.hibernate.mapping.Value;
-import org.hibernate.tool.hbm2x.visitor.DefaultValueVisitor;
+import org.hibernate.tool.internal.export.common.DefaultValueVisitor;
 import org.hibernate.tool.internal.util.NameConverter;
 import org.hibernate.tuple.GenerationTiming;
